solver: Be more careful with initialization state
We should log cases when we try to use the SimplexSolver API with an
uninitialized solver.

For SimplexSolvers that are part of a GObject, the structure gets memset
to zeros by the GObject instantiation, but if we put them on the stack
we're going to need an explicit variable initialization macro.