layout: Handle allocation and preferred size
We use an edit variable during allocation, to suggest the value of the
layout's width and height in order to solve the attributes of each
child.

For the preferred size of the children, we use a constraint on the
minimum size, to be greater than (or equal to) the LayoutChild's child.