examples: Restructure the constraints
We need more constraints to allow the solver to correctly interpret the
layout.

The list of constraints was generated via autolayout.js from the visual
format. Given the nature of the constraints, we're probably going to
need to change the ConstraintLayout API from:

  pack(child, constraints)
  pack(child, constraints)
  pack(child, constraints)

to something more in line with existing constraint-based layouts, like:

  add(child)
  add(child)
  add(child)
  set_constraints(constraints)