solver: Add unique id to variables
A great deal of the Cassowary simplex solver implementation relies on
implicit ordering in the variable maps inside expressions; the ordering
of the variables influences the order of substitutions, which ends up
influencing how the simplex gets solved.

Various Cassowary implementations either replace the hashing function
and the Map data type in order to achieve a stable iteration order; or,
like the case of the Python implementation, rely on the implicit order
of a Set data type to be stable.

If we use an ever-incrementing identifier for each Variable, and use
a sortable data type, we should be able to replicate the same behaviour
ourselves.