solver: Turn infeasible_rows in a Vec
We cannot iterate over infeasible rows during the optimization of the
dual of the simplex, as the pivot() operation at the end may add or
remove infeasible rows.

Instead, we turn the infeasible rows into a vector from a map, and
iterate over its length while popping the last item. We can also
simplify the memory management, since infeasible rows are owned by
something else in the solver while they are in flight.