vfl-parser: Increase conformance to VFL
We should implement the Apple VFL behaviour more closely, where
possible. For instance, we currently parse `viewName.attrName`, which is
not part of VFL but of the extended VFL syntax, which is a custom super
set of VFL only recognised by autolayout.js.

Additionally, the parser should take a dictionary of valid views and
metrics, to perform validation on those tokens.