Move constraint from description method to utility function
Instead of adding a new method to EmeusConstraintLayout we should follow
what Apple and autolayout.js do, and have a convenience global function
that takes an array of VFL definitions, a dictionary of views, and a
dictionary of metrics, and generates EmeusConstraint instances that can
then be used in an EmeusConstraintLayout.

This global function allows programmatic substitutions of view names
with appropriate widgets, as well as the programmatic substitution of
symbolic names with numeric constants. Additionally, it allows the
definition of the "default spacing" values in both directions.