Add `configure` to satisfy build-api
The build-api:

  https://github.com/cgwalters/build-api

is an attempt at standardising an API for projects running under build
tools and build bots. By conforming to this API we minimize the amount
of effort needed to import, built, and install a project into an
automated build environment — as well as minimize the amount of ad hoc
code that has to be written inside those automated build environments.

In order to achieve it, we need a simple `configure` shell script that
invokes `meson` for us, with a set of known arguments:

 * `--prefix`, for the prefix of the installation
 * `--libdir`, for the path used to install libraries and modules

The script will also generate a small Makefile that will just invoke
`ninja` using the appropriate targets:

 * `all`, for building the project
 * `install`, for installing the project (with support for `DESTDIR`)