Initial import of the Constraints Editor
The Editor is a small utility that can be used to visualize, edit, and
save constraints from VFL descriptions.

This commit mostly introduces the skeleton of the application, plus the
whole build modifications to ensure that we have a binary and a desktop
file to launch it, including potential translations.