layout: Implement a better preferred size request
Currently, the preferred size of a constraint layout relies on the
existing state of the solver; this means that, once an allocation
occurs, the layout will attempt to maintain that size as its own
preferred size.

If we remove children and constraints, this may lead to non obvious
behaviour, like a layout that cannot shrink any more.

Since we want the preferred size of the widget to be the preferred size
of the layout at rest, with no stay constants imposed on it, we need to
forcibly reset layout's size with a low priority stay constant, so that
querying the size will yield the preferred value.