Clean up the versioning macros
The versioning macros are not entirely correct, and do not allow for an
API-breaking development step.

While the API is still in flux, we should use a 9x minor version,
similarly to how GTK+ and other libraries approach the same issue; for
all intent and purposes, API introduced during the 9x cycle(s) belongs
to the next major API version.