build: Set the appropriate versioning for the shared library
We use the same versioning scheme that GLib and GTK+ use, with the API
version as part of the file name, and the minor and micro version as
part of the soname. This allows parallel installation based on the API
major version, with ABI-compatible updates within the same API version.