solver: Clean up refcounting and plug leaks
Reference counting is all over the place, because various operations on
the tableau of expressions can temporarily drop acquire and release
references on variables; to avoid invalid reads and writes, we need to
ensure that those variables survive optimization and pivoting passes.

The major culprit of leaking references is adding a new constraint with
an artificial variable, as it adds temporary rows and then calls
remove_row() to drop them; simplex_solver_remove_row() returns the
removed expression with its reference count intact, as that's what we
need when pivoting the simplex, which means we need an additional
reference count release when done.

In other cases, we create additional variables and expressions
internally and we need to transfer their ownership to the tableau.

We *could* solve these issues by using a floating reference counting
scheme, but that would come with its own headaches.

After a fairly exhaustive session of leak hunting with Valgrind, the
solver is now leak free.