layout: don't assume target object is a child
In create_child_constraint we assume that the source object
must be a child widget/layout when is not NULL but that is
not always the case. It can be the actual layout as well
and it also be a widget that doesn't belong to the layout.

Add extra checks to determine if the source object is the
actual layout and that belongs to the actual layout.